Assignment with arithmetic operators use same code as operations
